+++ pcap-convert/app/src/PcapConvertStats.cpp
@@ -14,13 +14,13 @@ PcapConvertStats::~PcapConvertStats()
 
 // Business logic
 
-void PcapConvertStats::recordWrittenPacket(uint32_t value)
+void PcapConvertStats::recordWrittenPacket(size_t value)
 {
     ++writtenPackets;
     bytesOut += value;
 }
 
-void PcapConvertStats::recordDroppedPacket(uint32_t value)
+void PcapConvertStats::recordDroppedPacket(size_t value)
 {
     ++droppedPackets;
     bytesDropped += value;
